requirements.txt
asgiref==3.3.4
cairocffi==1.2.0
CairoSVG==2.5.2
cffi==1.14.5
cssselect2==0.4.1
defusedxml==0.7.1
Django==3.2
django-crispy-forms==1.11.2
django-environ==0.4.5
django-filter==2.4.0
django-taggit==1.4.0
django-weasyprint==1.1.0
django-widget-tweaks==1.4.8
html5lib==1.1
Pillow==8.2.0
pkg-resources==0.0.0
psycopg2-binary==2.8.6
pycparser==2.20
pyphen==0.11.0
pytz==2021.1
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
tinycss2==1.1.0
WeasyPrint==52.5
webencodings==0.5.1
